+++ examples/softcoded-config/db/versions/1/schema.sql
@@ -1,0 +1,40 @@
+--  Sample employee database
+--  See changelog table for details
+--  Copyright (C) 2007,2008, MySQL AB
+--
+--  Original data created by Fusheng Wang and Carlo Zaniolo
+--  http://www.cs.aau.dk/TimeCenter/software.htm
+--  http://www.cs.aau.dk/TimeCenter/Data/employeeTemporalDataSet.zip
+--
+--  Current schema by Giuseppe Maxia
+--  Data conversion from XML to relational by Patrick Crews
+--
+-- This work is licensed under the
+-- Creative Commons Attribution-Share Alike 3.0 Unported License.
+-- To view a copy of this license, visit
+-- http://creativecommons.org/licenses/by-sa/3.0/ or send a letter to
+-- Creative Commons, 171 Second Street, Suite 300, San Francisco,
+-- California, 94105, USA.
+--
+--  DISCLAIMER
+--  To the best of our knowledge, this data is fabricated, and
+--  it does not correspond to real people.
+--  Any similarity to existing people is purely coincidental.
+--
+
+CREATE TABLE employees (
+    emp_no      INT             NOT NULL,
+    birth_date  DATE            NOT NULL,
+    first_name  VARCHAR(14)     NOT NULL,
+    last_name   VARCHAR(16)     NOT NULL,
+    gender      ENUM ('M','F')  NOT NULL,
+    hire_date   DATE            NOT NULL,
+    PRIMARY KEY (emp_no)
+);
+
+CREATE TABLE departments (
+    dept_no     CHAR(4)         NOT NULL,
+    dept_name   VARCHAR(40)     NOT NULL,
+    PRIMARY KEY (dept_no),
+    UNIQUE  KEY (dept_name)
+);