+++ src/Smrtr/MysqlVersionControl/Command/Status.php
@@ -5,19 +5,18 @@ namespace Smrtr\MysqlVersionControl\Command;
 use Smrtr\MysqlVersionControl\Command\Parameters\CommonParametersTrait;
 use Smrtr\MysqlVersionControl\Command\Parameters\ComposerParams;
 use Smrtr\MysqlVersionControl\Helper\Configuration;
-use Smrtr\MysqlVersionControl\Receiver\Up as UpReceiver;
+use Smrtr\MysqlVersionControl\Receiver\Status as StatusReceiver;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
- * Class Up is a symfony console command that updates the database to the latest version.
- * 
+ * Class Status is a symfony console command that reports the status of the given environment.
+ *
  * @package Smrtr\MysqlVersionControl\Command
  * @author Joe Green <joe.green@smrtr.co.uk>
  */
-class Up extends Command
+class Status extends Command
 {
     use CommonParametersTrait;
 
@@ -29,26 +28,13 @@ class Up extends Command
         // Parameters
         $this
             ->addEnvironmentArgument()
-            ->addMysqlBinArgument()
             ->addGlobalOptions()
-            ->addOption(
-                'no-schema',
-                null,
-                InputOption::VALUE_NONE,
-                'Skip execution of the schema files'
-            )
-            ->addOption(
-                'install-provisional-version',
-                null,
-                InputOption::VALUE_NONE,
-                'Install a provisional version which may still be in development and is not final.'
-            )
         ;
 
         // Name & description
         $this
-            ->setName("up")
-            ->setDescription('Install the latest database versions on the given environment')
+            ->setName("status")
+            ->setDescription('Reports the status of the database for the given environment')
         ;
     }
 
@@ -66,15 +52,12 @@ class Up extends Command
         $composerParams->applyComposerParams($this, $input);
         Configuration::applyConsoleConfigurationOptions($input);
 
-        $receiver = new UpReceiver;
+        $receiver = new StatusReceiver;
         return $receiver->execute(
             $input,
             $output,
             $input->getArgument('env'),
-            $input->getArgument('mysql-bin'),
             $input->getOption('versions-path'),
-            $input->getOption('no-schema'),
-            $input->getOption('install-provisional-version'),
             $input->getOption('provisional-version')
         );
     }