+++ CHANGELOG.md
@@ -3,10 +3,24 @@ Changelog
 
 ## 1.3.0 (wip)
 
+### General Improvements
+
+ - Added a new command: `status`
+ - Added composer parameter loading to CLI commands
  - Bundle all commands into single script smyver.php
- - Refactor: implemented the command pattern
- - Refactor: implemented the configuration adapter pattern with Ini and PhpArray adapters
  - Added port configuration option
+ - Refactor: implemented the command pattern
+ - Refactor: implemented the configuration adapter pattern
+
+### Configuration adapters:
+
+ - Ini (the default adapter)
+ - PhpArray
+ - PhpFile
+
+### Up command:
+
+ - Option `--provisional-version` is relative to project root if not an absolute path
 
 ## 1.2.0